backend/requirements.txt
flask==2.3.3
flask-cors==4.0.2
requests==2.32.0
python-dotenv==1.0.0
gunicorn==23.0.0
SQLAlchemy==2.0.27
pymysql==1.1.1
cryptography==44.0.1
psutil==5.9.8
pymetasploit3==1.0.3
werkzeug==3.0.6
uuid==1.30
pytz==2023.3 